Safari-609.1.11.1/LayoutTests/accessibility/gtk/entry-and-password.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN"> 
<html> 
<head>
<script src="../../resources/js-test-pre.js"></script> 
<script src="../../resources/accessibility-helper.js"></script>
</head> 
<body> 
 
<form>
<input type='text' name='entry' value='123' />
<input type='password' name='passwordEntry' value='123' />
</form>

<div id="stopElement">End of test</div>

<pre id="tree"></pre>
 
<p id="description"></p> 
<div id="console"></div> 
 
<script>
    description("This tests that the text associated with entry and password fields is correct.");

    if (window.accessibilityController) {
        // Build the accessibility tree up until 'End of test' is encountered.
        window.stopElement = accessibilityController.accessibleElementById("stopElement");

        document.body.focus();
        dumpAccessibilityTree(accessibilityController.focusedElement, stopElement, 0);
    }

</script>
<script src="../../resources/js-test-post.js"></script>
</body> 
</html> 
