LayoutTests/accessibility/gtk/entry-and-password-expected.txt


End of test
AXRole: AXWebArea
    AXRole: AXForm AXValue: <obj> <obj>
        AXRole: AXTextField AXValue: 123
        AXRole: AXPasswordField AXValue: •••
    AXRole: AXSection AXValue: End of test
This tests that the text associated with entry and password fields is correct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
